+++ src/types/schema.d.ts
@@ -348,6 +348,26 @@ export interface paths {
     patch?: never;
     trace?: never;
   };
+  "/proposals/{proposalReference}/shipments": {
+    parameters: {
+      query?: never;
+      header?: never;
+      path?: never;
+      cookie?: never;
+    };
+    /**
+     * Get Proposal Shipments
+     * @description Get shipments in proposal
+     */
+    get: operations["get_proposal_shipments_proposals__proposalReference__shipments_get"];
+    put?: never;
+    post?: never;
+    delete?: never;
+    options?: never;
+    head?: never;
+    patch?: never;
+    trace?: never;
+  };
   "/proposals/{proposalReference}/sessions/{visitNumber}/assign-data-collection-groups": {
     parameters: {
       query?: never;
@@ -544,6 +564,26 @@ export interface paths {
     patch?: never;
     trace?: never;
   };
+  "/sessions": {
+    parameters: {
+      query?: never;
+      header?: never;
+      path?: never;
+      cookie?: never;
+    };
+    /**
+     * Get Sessions
+     * @description Get sessions a user can view (wrapper for Expeye endpoint)
+     */
+    get: operations["get_sessions_sessions_get"];
+    put?: never;
+    post?: never;
+    delete?: never;
+    options?: never;
+    head?: never;
+    patch?: never;
+    trace?: never;
+  };
 }
 export type webhooks = Record<string, never>;
 export interface components {
@@ -557,7 +597,9 @@ export interface components {
       /** Capacity */
       capacity?: number | null;
       /** Details */
-      details?: Record<string, never> | null;
+      details?: {
+        [key: string]: unknown;
+      } | null;
       /** Location */
       location?: number | null;
       /**
@@ -593,7 +635,9 @@ export interface components {
       /** Capacity */
       capacity?: number | null;
       /** Details */
-      details?: Record<string, never> | null;
+      details?: {
+        [key: string]: unknown;
+      } | null;
       /** Location */
       location?: number | null;
       /**
@@ -657,7 +701,9 @@ export interface components {
       /** Capacity */
       capacity?: number | null;
       /** Details */
-      details?: Record<string, never> | null;
+      details?: {
+        [key: string]: unknown;
+      } | null;
       /** Location */
       location?: number | null;
       /**
@@ -692,7 +738,9 @@ export interface components {
       /** Sublocation */
       subLocation?: number | null;
       /** Details */
-      details?: Record<string, never> | null;
+      details?: {
+        [key: string]: unknown;
+      } | null;
       /** Comments */
       comments?: string | null;
       /**
@@ -710,7 +758,9 @@ export interface components {
     /** OptionalTopLevelContainer */
     OptionalTopLevelContainer: {
       /** Details */
-      details?: Record<string, never> | null;
+      details?: {
+        [key: string]: unknown;
+      } | null;
       /** Comments */
       comments?: string | null;
       /**
@@ -758,6 +808,17 @@ export interface components {
       /** Limit */
       limit: number;
     };
+    /** Paged[SessionOut] */
+    Paged_SessionOut_: {
+      /** Items */
+      items: components["schemas"]["SessionOut"][];
+      /** Total */
+      total: number;
+      /** Page */
+      page: number;
+      /** Limit */
+      limit: number;
+    };
     /** Paged[ShipmentOut] */
     Paged_ShipmentOut_: {
       /** Items */
@@ -783,12 +844,16 @@ export interface components {
     /** PreSessionIn */
     PreSessionIn: {
       /** Details */
-      details?: Record<string, never> | null;
+      details?: {
+        [key: string]: unknown;
+      } | null;
     };
     /** PreSessionOut */
     PreSessionOut: {
       /** Details */
-      details?: Record<string, never> | null;
+      details?: {
+        [key: string]: unknown;
+      } | null;
       /**
        * Islocked
        * @default false
@@ -804,7 +869,9 @@ export interface components {
       /** Sublocation */
       subLocation?: number | null;
       /** Details */
-      details?: Record<string, never> | null;
+      details?: {
+        [key: string]: unknown;
+      } | null;
       /** Comments */
       comments?: string | null;
       /**
@@ -833,7 +900,9 @@ export interface components {
       /** Sublocation */
       subLocation?: number | null;
       /** Details */
-      details?: Record<string, any> | null;
+      details?: {
+        [key: string]: unknown;
+      } | null;
       /** Comments */
       comments?: string | null;
       /**
@@ -862,6 +931,65 @@ export interface components {
       /** Externalid */
       externalId?: number | null;
     };
+    /** SessionOut */
+    SessionOut: {
+      /** Beamlinesetupid */
+      beamLineSetupId?: number | null;
+      /** Beamcalendarid */
+      beamCalendarId?: number | null;
+      /** Startdate */
+      startDate?: string | null;
+      /** Enddate */
+      endDate?: string | null;
+      /** Beamlinename */
+      beamLineName?: string | null;
+      /** Scheduled */
+      scheduled?: number | null;
+      /** Nbshifts */
+      nbShifts?: number | null;
+      /** Comments */
+      comments?: string | null;
+      /** Visitnumber */
+      visitNumber?: number | null;
+      /**
+       * Usedflag
+       * @description Indicates if session has Datacollections or XFE or EnergyScans attached
+       */
+      usedFlag?: number | null;
+      /**
+       * Lastupdate
+       * @description Last update timestamp: by default the end of the session, the last collect
+       */
+      lastUpdate?: string | null;
+      /** Parentproposal */
+      parentProposal?: string | null;
+      /**
+       * Proposalid
+       * @description Proposal ID
+       */
+      proposalId: number;
+      /**
+       * Sessionid
+       * @description Session ID
+       */
+      sessionId: number;
+      /** Beamlineoperator */
+      beamLineOperator?: string[] | null;
+      /**
+       * Bltimestamp
+       * Format: date-time
+       */
+      bltimeStamp: string;
+      /** Purgedprocesseddata */
+      purgedProcessedData: boolean;
+      /**
+       * Archived
+       * @description The data for the session is archived and no longer available on disk
+       */
+      archived: number;
+      /** Collectiongroups */
+      collectionGroups?: number | null;
+    };
     /** ShipmentChildren */
     ShipmentChildren: {
       /** Id */
@@ -871,7 +999,9 @@ export interface components {
       /** Children */
       children: components["schemas"]["GenericItem"][];
       /** Data */
-      data: Record<string, never>;
+      data: {
+        [key: string]: unknown;
+      };
     };
     /** ShipmentIn */
     ShipmentIn: {
@@ -944,7 +1074,9 @@ export interface components {
     /** TopLevelContainerIn */
     TopLevelContainerIn: {
       /** Details */
-      details?: Record<string, never> | null;
+      details?: {
+        [key: string]: unknown;
+      } | null;
       /** Comments */
       comments?: string | null;
       /**
@@ -965,7 +1097,9 @@ export interface components {
     /** TopLevelContainerOut */
     TopLevelContainerOut: {
       /** Details */
-      details?: Record<string, never> | null;
+      details?: {
+        [key: string]: unknown;
+      } | null;
       /** Comments */
       comments?: string | null;
       /**
@@ -979,11 +1113,8 @@ export interface components {
       type: string;
       /** Externalid */
       externalId?: number | null;
-      /**
-       * Barcode
-       * Format: uuid
-       */
-      barCode: string;
+      /** Barcode */
+      barCode?: string | null;
       /** History */
       history?: components["schemas"]["TopLevelContainerHistory"][] | null;
     };
@@ -1734,6 +1865,42 @@ export interface operations {
       };
     };
   };
+  get_proposal_shipments_proposals__proposalReference__shipments_get: {
+    parameters: {
+      query?: {
+        /** @description Page number/Results to skip. Negative numbers count backwards from the last page */
+        page?: number;
+        /** @description Number of results to show */
+        limit?: number;
+      };
+      header?: never;
+      path: {
+        proposalReference: string;
+      };
+      cookie?: never;
+    };
+    requestBody?: never;
+    responses: {
+      /** @description Successful Response */
+      200: {
+        headers: {
+          [name: string]: unknown;
+        };
+        content: {
+          "application/json": unknown;
+        };
+      };
+      /** @description Validation Error */
+      422: {
+        headers: {
+          [name: string]: unknown;
+        };
+        content: {
+          "application/json": components["schemas"]["HTTPValidationError"];
+        };
+      };
+    };
+  };
   assign_dcg_in_sublocation_proposals__proposalReference__sessions__visitNumber__assign_data_collection_groups_post: {
     parameters: {
       query?: never;
@@ -2145,6 +2312,41 @@ export interface operations {
         };
         content: {
           "application/json": components["schemas"]["ShipmentChildren"];
+        };
+      };
+      /** @description Validation Error */
+      422: {
+        headers: {
+          [name: string]: unknown;
+        };
+        content: {
+          "application/json": components["schemas"]["HTTPValidationError"];
+        };
+      };
+    };
+  };
+  get_sessions_sessions_get: {
+    parameters: {
+      query?: {
+        minEndDate?: string | null;
+        /** @description Page number/Results to skip. Negative numbers count backwards from the last page */
+        page?: number;
+        /** @description Number of results to show */
+        limit?: number;
+      };
+      header?: never;
+      path?: never;
+      cookie?: never;
+    };
+    requestBody?: never;
+    responses: {
+      /** @description Successful Response */
+      200: {
+        headers: {
+          [name: string]: unknown;
+        };
+        content: {
+          "application/json": components["schemas"]["Paged_SessionOut_"];
         };
       };
       /** @description Validation Error */